+++ src/store/store.ts
@@ -1,9 +1,11 @@
-import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
+import { configureStore, ThunkAction, Action } from "@reduxjs/toolkit";
+import counterReducer from "../features/counter/counterSlice";
+import cryptoCoinsReducer from "../features/cryptoCoins/cryptoCoinsSlice";
 
 export const store = configureStore({
   reducer: {
     counter: counterReducer,
+    cryptoCoins: cryptoCoinsReducer,
   },
 });
